+++ test/oap3/format/OAR016/fail-example.js
@@ -24,7 +24,7 @@ module.exports = {
                       "properties": {
                         "productId": {
                           "type": "integer",
-                          "format": "big"
+                          "format": "double"
                         }
                       }
                     }